pyproject.toml
[tool.poetry]
name = "mailshot-site"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
django-celery-beat = "^2.6.0"
django-formtools = "^2.5.1"
pillow = "^10.3.0"
psycopg2 = "^2.9.9"
redis = "^5.0.4"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
